[#11] Refactor Home screen to support display on large screen width (tablet)
