+++ app/src/main/java/co/nimblehq/compose/ecommerce/ui/bottomnavigationbar/Navigation.kt
@@ -10,17 +10,18 @@ import co.nimblehq.compose.ecommerce.ui.screen.account.AccountScreen
 import co.nimblehq.compose.ecommerce.ui.screen.home.HomeScreen
 import co.nimblehq.compose.ecommerce.ui.screen.product.ProductScreen
 import co.nimblehq.compose.ecommerce.ui.screen.search.SearchScreen
+import co.nimblehq.compose.ecommerce.utils.WindowSize
 
 @ExperimentalFoundationApi
 @OptIn(ExperimentalMaterialApi::class)
 @Composable
-fun Navigation(navHostController: NavHostController) {
+fun Navigation(navHostController: NavHostController, windowSize: WindowSize) {
     NavHost(
         navController = navHostController,
         startDestination = NavigationItem.Home.route
     ) {
         composable(NavigationItem.Home.route) {
-            HomeScreen()
+            HomeScreen(windowSize)
         }
         composable(NavigationItem.Search.route) {
             SearchScreen()
